hwsim: Add support for HWSIM_CMD_GET_RADIO
Add a new option to the hwsim binary to fetch all or a given hwsim radio.
Print out radio information which is currently returned in the reply to
user space.

Add new hwsim attributes to the enum and tweak getopt to behave properly
with or without giving the radio id on the command line.