test-runner: Use '-cpu host' in kvm mode
When KVM is available, use the -cpu host option.  This is useful on
distributions that might use architecture specific optimizations, which
might not be available on the default machine being emulated by qemu.

For example, a Gentoo box compiled with "-march=core-avx2" would result
in several binaries not executing properly inside the qemu environment
due to instruction set mismatches.